requirements.txt
google-search-results==2.4.2
aiogram==3.2.0
pyyaml
python-dotenv